+++ waffle_utils/dataset/__init__.py
@@ -1,17 +1,4 @@
-import enum
-
 from .dataset import Dataset
+from .format import Format
 
-__all__ = ["Dataset"]
-
-
-# enums
-class Format(enum.Flag):
-    YOLO_CLASSIFICATION = enum.auto()
-    YOLO_DETECTION = enum.auto()
-    YOLO_SEGMENTATION = enum.auto()
-
-    COCO_DETECTION = enum.auto()
-
-
-# formats
+__all__ = ["Dataset", "Format"]
